Handle Gospel specifications
Gospel specifications are in the list of attributes attached to a value
or type declaration. There may be several gospel attributes attached to
one declaration. We build the corresponding documentation attributes and
append them to the list of attributes.

TODO: test behaviour with multiple Gospel attributes attached to one
declaration.